requirements.txt
Django==3.2.12
django_celery_beat==2.3.0
django_filter==21.1
django_shortuuidfield~=0.1.3
djangorestframework~=3.14.0
djangorestframework_simplejwt==5.1.0
PyMySQL~=1.1.0
celery==5.2.3
django-celery-results==2.4.0
django-cors-headers==3.11.0
django-simple-history==3.0.0
drf-yasg
coreapi
django_redis
